requirements.txt
streamlit
pandas
Pillow
boto3
json
time
requests
